requirements.txt
asgiref==3.4.1
astroid==2.6.2
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==1.1.0
Django==3.2.5
django-allauth==0.44.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-yasg==1.17.1
drf-yasg2==1.19.4
idna==2.10
inflection==0.5.1
isort==5.9.2
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.9.3
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
